requirements.txt
fastapi>=0.115.2
torch==2.4.1
torchvision==0.19.1
numpy==2.1.2
pillow==10.4.0
python-multipart==0.0.12
requests==2.32.3
mangum==0.17.0